クレジットカードを検出したら前画面に戻るように対応
