+++ Sample/MlKitSample/app/build.gradle
@@ -36,6 +36,7 @@ dependencies {
     implementation 'androidx.appcompat:appcompat:1.1.0'
     implementation 'androidx.core:core-ktx:1.2.0'
     implementation 'androidx.constraintlayout:constraintlayout:1.1.3'
+    implementation 'com.google.android.material:material:1.1.0'
 
     // for sample
     implementation 'com.google.android.gms:play-services-mlkit-text-recognition:16.1.0'
