+++ Sample/MlKitSample/app/src/main/AndroidManifest.xml
@@ -15,10 +15,10 @@
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
-
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity android:name=".CameraActivity" />
     </application>
 
 </manifest>
