+++ Sample/MlKitSample/app/src/main/java/com/github/mag0716/mlkitsample/CreditCard.kt
@@ -1,0 +1,8 @@
+package com.github.mag0716.mlkitsample
+
+import java.io.Serializable
+
+data class CreditCard(
+    val number: String,
+    val expiredDate: String
+) : Serializable
